i do not know what i have: for i have written unholy code: it must not stray further from sanity and the clarity of mind

+++ src/main/java/me/urkkiz/util/TimeManager.java
@@ -3,7 +3,7 @@ package me.urkkiz.util;
 import java.util.ArrayList;
 
 public class TimeManager {
-    public static ArrayList<Long> Timers=new ArrayList<>();
+    public static ArrayList<Float> Timers=new ArrayList<>();
     public static float frameEndTime;
     public static void CalcFrameEndTime(){
         frameEndTime=((System.nanoTime()-frameEndTime));
